refactor: 💡 rename resource into resources
(I had an error from inside ffspec, seems like there is a name conflict
- weird!)
