+++ libs/libcommon/src/libcommon/resources.py
