+++ workers/datasets_based/src/datasets_based/asset.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import List, Tuple, TypedDict
 
 import soundfile  # type:ignore
-from libcommon.resource import StrPath
+from libcommon.resources import StrPath
 from numpy import ndarray  # type:ignore
 from PIL import Image  # type: ignore
 from pydub import AudioSegment  # type:ignore
