+++ workers/datasets_based/tests/workers/test_split_names.py
@@ -8,11 +8,11 @@ import pytest
 from libcommon.exceptions import CustomError
 from libcommon.processing_graph import ProcessingStep
 from libcommon.queue import Priority
-from libcommon.resource import CacheDatabaseResource, QueueDatabaseResource
+from libcommon.resources import CacheDatabaseResource, QueueDatabaseResource
 from libcommon.simple_cache import DoesNotExist, get_response
 
 from datasets_based.config import AppConfig
-from datasets_based.resource import LibrariesResource
+from datasets_based.resources import LibrariesResource
 from datasets_based.workers.split_names import SplitNamesWorker
 
 from ..fixtures.hub import HubDatasets, get_default_config_split
